Refactor filter_statuses and add logging
